reorganize into namespaces
