+++ server/run.py
@@ -1,0 +1,6 @@
+from app import create_app
+from config import DevConfig
+
+if __name__ == '__main__':
+    app = create_app(DevConfig)
+    app.run()
